# target/enunciate/gwt-json-overlay/src/org/example/prispevky/entity/Comment.java
/**
 * 
 *
 * Generated by <a href="http://enunciate.webcohesion.com">Enunciate</a>.
 */
package org.example.prispevky.entity;

/**
 * (no documentation provided)
 */
public class Comment extends com.google.gwt.core.client.JavaScriptObject {

  protected Comment() {
  }

  /**
   * Create an instance of Comment from JSON text.
   */
  public static native Comment fromJson(String json) /*-{
    return eval('(' + json + ')');
  }-*/;
}