# target/enunciate/java-json-client/src/org/example/prispevky/entity/Comment.java
/**
 * 
 *
 * Generated by <a href="http://enunciate.webcohesion.com">Enunciate</a>.
 */
package org.example.prispevky.entity;

import com.fasterxml.jackson.annotation.*;

/**
 * (no documentation provided)
 */
@JsonInclude( JsonInclude.Include.NON_NULL )
@JsonIgnoreProperties( ignoreUnknown = true )
public class Comment implements java.io.Serializable {

}