# target/enunciate/java-xml-client/src/org/example/prispevky/entity/Comment.java
/**
 * 
 *
 * Generated by <a href="http://enunciate.webcohesion.com">Enunciate</a>.
 */
package org.example.prispevky.entity;

/**
 * (no documentation provided)
 */
@jakarta.xml.bind.annotation.XmlType (
  name = "comment",
  namespace = ""
)
public class Comment implements java.io.Serializable {


}